Log file is optional for robocopy sync

+++ bootstrap.ps1
@@ -29,37 +29,45 @@ function doIt() {
 
   # sync the files
   $files = $filesToBuild -join " "
-  sync $buildFolder $files "robocopyLog1.txt"
+  sync $buildFolder $home $files "robocopyLog1.txt"
 
   # sync other files
-  sync $rootFolder $otherFiles "robocopyLog2.txt"
+  sync $rootFolder $home $otherFiles "robocopyLog2.txt"
 }
 
-function sync($sourceFolder, $files, $logfile) {
-  $destinationFolder = $home
+function sync($sourceFolder, $destinationFolder, $files, $logfile=$null) {
   # We do not use /PURGE (or /MIR) because then robocopy
   # takes too long to proceed, and we don't want to delete
   # anything in ~ that doesn't exist in this dotfiles repo.
   $program="robocopy"
   # /COPYALL: Copy all file info (timestamp, owner, ACL's etc)
+  # /S: Copy subdirectories (not empty ones)
   # /LEV:1: Do not any copy subfolders (only first level)
   # Note: We could use /PURGE instead of /MIR
-  $options="/COPYALL /LEV:1"
-  $loggingOptions="/V /TEE /NC /NS /NDL /LOG:$logfile"
+  $options="/COPYALL"
+  if ( $logfile -eq $null) {
+    $loggingOptions="/V /TEE /NC /NS /NDL"
+  }
+  else {
+    $loggingOptions="/V /TEE /NC /NS /NDL /LOG:$logfile"
+  }
   $retryOptions="/R:0 /W:0"
   $excludedFiles="README.md", "bootstrap.sh", "bootstrap.ps1", "LICENSE-MIT.txt"
   $excludedFilesOptions = "/XF $($excludedFiles -join ' /XF ')"
-  $excludedDirectories = ".git", ".vim", "bin", "init", "scripts"
+  $excludedDirectories = ".git", ".vim", "build", "docs", "init", "scripts"
   $excludedDirectoriesOptions = "/XD $($excludedDirectories -join ' /XD ')"
 
-  # Note: You can insert /L option to debug any pronlems
+  # Note: You can insert /L option to debug any problems
   # /L: List only, doesn't copy anything (-WhatIf in powershell)
 
   # Perform sync with robocopy tool
   Invoke-Expression "$program $sourceFolder $destinationFolder $files $options $loggingOptions $retryOptions"
 
   # Move logfile to ~
-  Invoke-Expression "mv -Force $(Join-Path $rootFolder $logfile) $destinationFolder"
+  if ( Test-Path (Join-Path $rootFolder $logfile) )
+  {
+    Invoke-Expression "mv -Force $(Join-Path $rootFolder $logfile) $destinationFolder"
+  }
 }
 
 if ( $args[0] -eq "--build" -or $args[0] -eq "-b" ) {
